@@ -9,80 +9,102 @@
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "webpack:server": "webpack --config webpack.server.config.js --progress --colors"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.0",
-    "@angular/cdk": "6.0.0",
-    "@angular/common": "6.0.0",
-    "@angular/compiler": "6.0.0",
-    "@angular/core": "6.0.0",
-    "@angular/flex-layout": "^6.0.0-beta.15",
-    "@angular/forms": "6.0.0",
-    "@angular/http": "6.0.0",
-    "@angular/material": "6.0.0",
-    "@angular/platform-browser": "6.0.0",
-    "@angular/platform-browser-dynamic": "6.0.0",
-    "@angular/router": "6.0.0",
-    "@fortawesome/angular-fontawesome": "^0.1.1",
-    "@fortawesome/fontawesome-svg-core": "^1.2.0",
-    "@fortawesome/free-solid-svg-icons": "^5.1.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.1.2",
+    "@angular/animations": "^7.1.4",
+    "@angular/cdk": "7.2.0",
+    "@angular/common": "7.1.4",
+    "@angular/compiler": "7.1.4",
+    "@angular/core": "7.1.4",
+    "@angular/fire": "^5.1.1",
+    "@angular/flex-layout": "^7.0.0-beta.22",
+    "@angular/forms": "7.1.4",
+    "@angular/http": "7.1.4",
+    "@angular/material": "7.2.0",
+    "@angular/platform-browser": "7.1.4",
+    "@angular/platform-browser-dynamic": "7.1.4",
+    "@angular/platform-server": "^7.1.4",
+    "@angular/router": "7.1.4",
+    "@fortawesome/angular-fontawesome": "^0.3.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.12",
+    "@fortawesome/free-solid-svg-icons": "^5.6.3",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.1",
+    "@nguniversal/module-map-ngfactory-loader": "^7.0.2",
+    "@types/lodash": "^4.14.119",
+    "angular-bootstrap-md": "^7.2.0",
     "angular-font-awesome": "^3.1.2",
     "angular-sanitize": "^1.7.5",
-    "angularfire2": "^5.0.0-rc.6",
+    "angularfire2": "5.1.1",
     "body-parser": "^1.18.3",
-    "bootstrap": "^4.1.3",
+    "bootstrap": "^4.2.1",
     "compass": "^0.1.1",
-    "core-js": "^2.4.1",
-    "express": "^4.16.3",
+    "core-js": "^2.6.1",
+    "emailjs-com": "^2.3.2",
+    "express": "^4.16.4",
     "express-handlebars": "^3.0.0",
-    "firebase": "^4.12.0",
+    "firebase": "^5.7.1",
     "font-awesome": "^4.7.0",
-    "foundation-sites": "^6.4.4-rc1",
-    "hexo": "^3.7.1",
+    "foundation-sites": "^6.5.1",
+    "hexo": "^3.8.0",
     "hoverifyBootnav": "^1.0.0",
-    "intersection-observer": "^0.5.0",
-    "ionicons": "^4.1.2",
+    "intersection-observer": "^0.5.1",
+    "ionicons": "^4.5.1",
     "isotope": "^1.0.0-alpha.3",
     "isotope-layout": "^3.0.6",
-    "jquery": "^3.3.1",
-    "json-server": "^0.12.2",
-    "ngx-bootstrap": "^2.0.3",
+    "json-server": "^0.14.0",
+    "lodash": "^4.17.11",
+    "lodash.filter": "^4.6.0",
+    "lodash.flatmap": "^4.5.0",
+    "lodash.flow": "^3.5.0",
+    "ng-back-to-top": "0.0.3",
+    "ng-recaptcha": "^4.2.1",
+    "ngx-bootstrap": "^3.1.3",
+    "ngx-foundation": "^1.0.6",
     "ngx-lazy-load-images": "^1.3.1",
-    "ngx-loading": "^3.0.1",
-    "ngx-scrollspy": "^1.2.0",
-    "ngx-spinner": "^6.1.2",
-    "nodemailer": "^4.6.7",
-    "npm": "^6.4.1",
-    "popper.js": "^1.14.3",
-    "rxjs": "^6.1.0",
-    "rxjs-compat": "^6.0.0",
+    "ngx-page-scroll": "^5.0.1",
+    "ngx-scrollspy": "^1.2.1",
+    "nodemailer": "^4.7.0",
+    "npm": "^6.5.0",
+    "popper.js": "^1.14.6",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "sal.js": "^0.5.0",
+    "sass-loader": "^7.1.0",
+    "scroll-out": "^2.2.4",
+    "slick-carousel": "^1.8.1",
+    "what-input": "^5.1.3",
+    "yarn": "^1.13.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "6.0.0",
-    "@angular/compiler-cli": "6.0.0",
-    "@angular/language-service": "6.0.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "^0.11.4",
+    "@angular/cli": "^7.1.4",
+    "@angular/compiler-cli": "7.1.4",
+    "@angular/language-service": "7.1.4",
+    "@types/jasmine": "~3.3.5",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/jquery": "^3.3.27",
+    "@types/node": "^10.12.18",
+    "codelyzer": "^4.5.0",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "jquery": "^3.3.1",
-    "karma": "~2.0.0",
+    "karma": "~3.1.4",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "motion-ui": "^1.2.3",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "2.7.2",
-    "@angular-devkit/build-angular": "~0.6.0"
+    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "motion-ui": "^2.0.3",
+    "ngx-parallax": "^3.0.0-rc.2",
+    "protractor": "^5.4.2",
+    "ts-loader": "^5.3.2",
+    "ts-node": "~7.0.1",
+    "tslint": "~5.12.0",
+    "typescript": "^3.1.6",
+    "webpack": "^4.28.2",
+    "webpack-cli": "^3.1.2"
   }
 }
