@@ -1,13 +1,14 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { IProject } from "../project";
-import { Observable } from "rxjs/Observable";
+import { Observable } from 'rxjs';
+
+import { IProject } from '../project';
 
 @Injectable()
 export class ProjectService {
 
-  private _url: string = "http://www.marvinrusinek.com/assets/data/project.json";
-  //private _url: string = "../assets/data/project.json";
+  private _url = 'http://www.marvinrusinek.com/assets/data/project.json';
+  // private _url = '../assets/data/project.json';
 
   constructor(private http: HttpClient) {}
 
