@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
-import { MatButtonModule } from '@angular/material';
 
 @Component({
   selector: 'app-portfolio-hero',
   templateUrl: './hero.component.html',
   styleUrls: ['./hero.component.scss']
 })
+
 export class HeroComponent {
-  constructor() {}
 }
