@@ -5,6 +5,6 @@ import { Component } from '@angular/core';
   templateUrl: './page-not-found.component.html',
   styleUrls: ['./page-not-found.component.css']
 })
+
 export class PageNotFoundComponent {
-  constructor() { }
 }
