@@ -1,14 +1,14 @@
 export interface IProject {
-  id: number,
-  category: string,
-  categoryClass: string,
-  categoryFullNames: string,
-  projectUrl: string,
-  codeUrl: string,
-  imagePath: string,
-  imageTitle: string,
-  title: string,
-  dateCompleted: string,
-  technologies: string,
-  description: string
+  id: number;
+  category: string;
+  categoryClass: string;
+  categoryFullNames: string;
+  projectUrl: string;
+  codeUrl: string;
+  imagePath: string;
+  imageTitle: string;
+  title: string;
+  dateCompleted: string;
+  technologies: string;
+  description: string;
 }
