@@ -1,9 +1,9 @@
 export const environment = {
   production: true,
   firebase: {
-    apiKey: "AIzaSyBhJEg2usnAjQ0_ybTIfIw3oGfFaTz6lYE",
-    authDomain: "marvin-rusinek-portfolio-web.firebaseapp.com",
-    databaseURL: "https://marvin-rusinek-portfolio-web.firebaseio.com",
-    storageBucket: "marvin-rusinek-portfolio-web.appspot.com"
+    apiKey: 'AIzaSyBhJEg2usnAjQ0_ybTIfIw3oGfFaTz6lYE',
+    authDomain: 'marvin-rusinek-portfolio-web.firebaseapp.com',
+    databaseURL: 'https://marvin-rusinek-portfolio-web.firebaseio.com',
+    storageBucket: 'marvin-rusinek-portfolio-web.appspot.com'
   }
 };
